+++ src/com/facebook/buck/jvm/kotlin/BUCK.autodeps
@@ -1,17 +1,16 @@
-#@# GENERATED FILE: DO NOT MODIFY e662c9159ac889a60b28d711f092a8618ecdaab5 #@#
+#@# GENERATED FILE: DO NOT MODIFY 3b47595dd70b40b437a765f15169477d7739b369 #@#
 {
   "kotlin" : {
     "deps" : [
-      "//src/com/facebook/buck/io:executable-finder",
       "//src/com/facebook/buck/jvm/common:common"
     ],
     "exported_deps" : [
-      "//src/com/facebook/buck/cli:config",
       "//src/com/facebook/buck/io:io",
       "//src/com/facebook/buck/jvm/core:suggestbuildrules",
       "//src/com/facebook/buck/jvm/java:rules",
       "//src/com/facebook/buck/jvm/java:steps",
       "//src/com/facebook/buck/jvm/java:support",
+      "//src/com/facebook/buck/jvm/kotlin/config:config",
       "//src/com/facebook/buck/model:model",
       "//src/com/facebook/buck/parser:rule_pattern",
       "//src/com/facebook/buck/rules:build_rule",
