+++ src/com/facebook/buck/jvm/kotlin/config/BUCK
@@ -1,0 +1,1 @@
+standard_java_library()
